fix: Use separate workflows for updating nixpkgs, pint-src
This fixes an issue where previously, the nixpkgs input would be updated
on a sunday, but if noone merged the PR before the workflow ran again
the next night, the nixpkgs update would get overridden with the next
pint-src update force-pushed by the workflow.

By using separate workflows, both are a lot simpler to read, and both
will open separate PRs.